requirements.txt
luigi
xrootd-client
